refactor: rename for uniformity
