+++ src/modules/Entities/PlayerEntity.js
