+++ src/modules/Entities/ScrollingBackgroundEntity.js
